requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
aiofiles==24.1.0
    # via unstructured-client
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
backoff==2.2.1
    # via unstructured
beautifulsoup4==4.13.4
    # via unstructured
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via unstructured
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   nltk
    #   python-oxmsg
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==45.0.5
    # via unstructured-client
dataclasses-json==0.6.7
    # via unstructured
emoji==2.14.1
    # via unstructured
et-xmlfile==2.0.0
    # via openpyxl
filetype==1.2.0
    # via unstructured
h11==0.16.0
    # via httpcore
html5lib==1.1
    # via unstructured
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via unstructured-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
joblib==1.5.1
    # via nltk
langdetect==1.0.9
    # via unstructured
lxml==6.0.0
    # via unstructured
marshmallow==3.26.1
    # via dataclasses-json
mypy-extensions==1.1.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via unstructured-client
nltk==3.9.1
    # via unstructured
numpy==2.3.1
    # via
    #   pandas
    #   unstructured
olefile==0.47
    # via python-oxmsg
openpyxl==3.1.5
    # via -r .\requirements.in
packaging==25.0
    # via marshmallow
pandas==2.3.1
    # via -r .\requirements.in
psutil==7.0.0
    # via unstructured
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r .\requirements.in
    #   unstructured-client
pydantic-core==2.33.2
    # via pydantic
pypdf==5.8.0
    # via
    #   -r .\requirements.in
    #   unstructured-client
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r .\requirements.in
python-iso639==2025.2.18
    # via unstructured
python-magic==0.4.27
    # via unstructured
python-oxmsg==0.0.2
    # via unstructured
pytz==2025.2
    # via pandas
rapidfuzz==3.13.0
    # via unstructured
regex==2024.11.6
    # via nltk
requests==2.32.4
    # via
    #   -r .\requirements.in
    #   requests-toolbelt
    #   unstructured
requests-toolbelt==1.0.0
    # via unstructured-client
six==1.17.0
    # via
    #   html5lib
    #   langdetect
    #   python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
tqdm==4.67.1
    # via
    #   nltk
    #   unstructured
typing-extensions==4.14.1
    # via
    #   anyio
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   python-oxmsg
    #   typing-inspect
    #   typing-inspection
    #   unstructured
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
unstructured==0.18.9
    # via -r .\requirements.in
unstructured-client==0.39.1
    # via unstructured
urllib3==2.5.0
    # via requests
webencodings==0.5.1
    # via html5lib
wrapt==1.17.2
    # via unstructured
xlsxwriter==3.2.5
    # via -r .\requirements.in
